Refactored DNS Functions
I did some work to make the functions handling DNS manipulation a
little more reusable and better named.

PandaCluster requires that the user have a public hosted zone
established for a root-domain.  But now, it can create a new subdomain
if it doesn’t exist or change it if it does.  It can detect the
difference now as well.

Example:
- User registers example.com. with AWS and sets up a hosted zone.
- specifying test.example.com as the cluster’s address will add that
record to Route53.
- if test.example.com is already assigned, PandaCluster will detect
that and make the appropriate changes.