Places Kick Server On Cluster
There still needs to be some work done on setting up Security Groups,
but this successfully gets a Docker container on the cluster, feeds it
the user’s AWS credentials and, and starts it up.

There was also some refactoring of the service files.  Things needed
during cluster-formation are in `src/formation-services`.  Everything
else is in `/src/launch` so it’s stored in a way that is isomorphic to
how it’s deployed.